+++ src/routes/teacher.rs
@@ -14,7 +14,7 @@ async fn table(
         tera.autoescape_on(vec![]);
 
         let (id, th_name) = TeachRec::split_id_and_name(user.id().unwrap());
-        let id = id.parse().map_or(id, |id: i16| format!("{:04}", id));
+        let id = id.parse().map_or(id, |id: i32| format!("{:04}", id));
 
         let opens = crate::routes::index::read_attendance_dir(id.as_str())("attendance/open")
             .unwrap_or(Vec::new()); // todo: report errors
